+++ projects/pastanaga-angular/src/lib/table/table-cell/table-cell.component.ts
@@ -31,6 +31,16 @@ export class TableCellComponent {
     set disabled(value: boolean) {
         this._disabled = coerceBooleanProperty(value);
     }
+
+    @Input()
+    get noWrap(): boolean {
+        return this._noWrap;
+    }
+    set noWrap(value: boolean) {
+        this._noWrap = coerceBooleanProperty(value);
+    }
+
+    private _noWrap = false;
     private _disabled = false;
     private _header = false;
     private _clickable = false;